fix: Code conflict resolution
- Entity 중복문제 해결
