+++ src/main/java/com/akatsuki/pioms/ask/repository/FranchiseOwnerRepository.java
@@ -1,6 +1,7 @@
 package com.akatsuki.pioms.ask.repository;
 
-import com.akatsuki.pioms.ask.entity.FranchiseOwnerEntity;
+
+import com.akatsuki.pioms.ask.aggregate.FranchiseOwnerEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface FranchiseOwnerRepository extends JpaRepository<FranchiseOwnerEntity, Integer> {
