+++ src/main/java/com/akatsuki/pioms/order/dto/OrderDTO.java
@@ -1,6 +1,6 @@
 package com.akatsuki.pioms.order.dto;
 
-import com.akatsuki.pioms.franchise.entity.FranchiseEntity;
+import com.akatsuki.pioms.franchise.aggregate.Franchise;
 import com.akatsuki.pioms.order.etc.ORDER_CONDITION;
 
 import java.time.LocalDateTime;
@@ -13,5 +13,5 @@ public class OrderDTO {
     private ORDER_CONDITION orderCondition;
     private String orderReason;
     private boolean orderStatus;
-    private FranchiseEntity franchise;
+    private Franchise franchise;
 }
