+++ src/main/java/com/akatsuki/pioms/order/entity/OrderEntity.java
@@ -1,14 +1,12 @@
 package com.akatsuki.pioms.order.entity;
 
 import com.akatsuki.pioms.exchange.entity.ExchangeEntity;
-import com.akatsuki.pioms.franchise.entity.FranchiseEntity;
+import com.akatsuki.pioms.franchise.aggregate.Franchise;
 import com.akatsuki.pioms.order.etc.ORDER_CONDITION;
-import com.akatsuki.pioms.order.vo.OrderVO;
 import jakarta.persistence.*;
 import lombok.*;
 
 import java.time.LocalDateTime;
-import java.util.Date;
 import java.util.List;
 
 @Entity
@@ -42,7 +40,7 @@ public class OrderEntity {
 
     @JoinColumn(name = "franchise_code")
     @ManyToOne
-    private FranchiseEntity franchise;
+    private Franchise franchise;
 
     @JoinColumn(name = "exchange_code")
     @OneToOne
